+++ lithops/serverless/backends/code_engine/entry_point.py
@@ -15,29 +15,27 @@
 #
 
 import os
+import sys
+import json
 import logging
 import pkgutil
-import sys
 from lithops.version import __version__
-from lithops.config import cloud_logging_config
+from lithops.config import default_logging_config
 from lithops.worker import function_handler
-from lithops.worker import function_invoker
 from lithops.storage import InternalStorage
 from lithops.config import JOBS_PREFIX
 from lithops.utils import sizeof_fmt
 from lithops.storage.utils import create_runtime_meta_key
 
 
-
-cloud_logging_config(logging.DEBUG)
-import json
 logger = logging.getLogger('__main__')
 
 
 def binary_to_dict(the_binary):
     jsn = ''.join(chr(int(x, 2)) for x in the_binary.split())
-    d = json.loads(jsn)  
+    d = json.loads(jsn)
     return d
+
 
 def runtime_packages(storage_config):
     logger.info("Extracting preinstalled Python modules...")
@@ -48,7 +46,7 @@ def runtime_packages(storage_config):
     runtime_meta['preinstalls'] = [entry for entry in sorted([[mod, is_pkg] for _, mod, is_pkg in mods])]
     python_version = sys.version_info
     runtime_meta['python_ver'] = str(python_version[0])+"."+str(python_version[1])
-    
+
     activation_id = storage_config['activation_id']
 
     status_key = create_runtime_meta_key(JOBS_PREFIX, activation_id)
@@ -60,15 +58,19 @@ def runtime_packages(storage_config):
 
 
 def main(action, payload_decoded):
-    logger.info ("Welcome to Lithops-Code-Engine entry point. Action {}".format(action))
-         
+    logger.info("Welcome to Lithops-Code-Engine entry point. Action {}".format(action))
+
     payload = binary_to_dict(payload_decoded)
+
+    log_level = payload['log_level']
+    default_logging_config(log_level)
+
     logger.info(payload)
     if (action == 'preinstals'):
         runtime_packages(payload)
         return {"Execution": "Finished"}
     job_index = os.environ['JOB_INDEX']
-    logger.info (" Action {}. Job Index {}".format(action, job_index))
+    logger.info("Action {}. Job Index {}".format(action, job_index))
     os.environ['__PW_ACTIVATION_ID'] = payload['activation_id']
     payload['JOB_INDEX'] = job_index
     if 'remote_invoker' in payload:
@@ -86,6 +88,7 @@ def main(action, payload_decoded):
         function_handler(payload)
 
     return {"Execution": "Finished"}
+
 
 if __name__ == '__main__':
     main(sys.argv[1:][0], sys.argv[1:][1])